+++ datamodel/openapi/openapi-api-sample/src/main/java/com/sap/cloud/sdk/datamodel/openapi/sample/model/Cola.java
@@ -140,6 +140,7 @@ public class Cola implements OneOf, OneOfWithDiscriminator, OneOfWithDiscriminat
     /**
      * Get the value of an unrecognizable property of this {@link Cola} instance.
      *
+     * @deprecated Use {@link #toMap()} instead.
      * @param name
      *            The name of the property
      * @return The value of the property
@@ -147,6 +148,7 @@ public class Cola implements OneOf, OneOfWithDiscriminator, OneOfWithDiscriminat
      *             If no property with the given name could be found.
      */
     @Nullable
+    @Deprecated
     public Object getCustomField( @Nonnull final String name )
         throws NoSuchElementException
     {
@@ -154,6 +156,23 @@ public class Cola implements OneOf, OneOfWithDiscriminator, OneOfWithDiscriminat
             throw new NoSuchElementException("Cola has no field with name '" + name + "'.");
         }
         return cloudSdkCustomFields.get(name);
+    }
+
+    /**
+     * Get the value of all properties of this {@link Cola} instance including unrecognized properties.
+     *
+     * @return The map of all properties
+     */
+    @JsonIgnore
+    @Nonnull
+    public Map<String, Object> toMap()
+    {
+        final Map<String, Object> declaredFields = new LinkedHashMap<>(cloudSdkCustomFields);
+        if( sodaType != null )
+            declaredFields.put("sodaType", sodaType);
+        if( caffeine != null )
+            declaredFields.put("caffeine", caffeine);
+        return declaredFields;
     }
 
     /**