+++ datamodel/openapi/openapi-api-sample/src/main/java/com/sap/cloud/sdk/datamodel/openapi/sample/model/Fanta.java
@@ -179,6 +179,7 @@ public class Fanta implements OneOf, OneOfWithDiscriminator, OneOfWithDiscrimina
     /**
      * Get the value of an unrecognizable property of this {@link Fanta} instance.
      *
+     * @deprecated Use {@link #toMap()} instead.
      * @param name
      *            The name of the property
      * @return The value of the property
@@ -186,6 +187,7 @@ public class Fanta implements OneOf, OneOfWithDiscriminator, OneOfWithDiscrimina
      *             If no property with the given name could be found.
      */
     @Nullable
+    @Deprecated
     public Object getCustomField( @Nonnull final String name )
         throws NoSuchElementException
     {
@@ -193,6 +195,25 @@ public class Fanta implements OneOf, OneOfWithDiscriminator, OneOfWithDiscrimina
             throw new NoSuchElementException("Fanta has no field with name '" + name + "'.");
         }
         return cloudSdkCustomFields.get(name);
+    }
+
+    /**
+     * Get the value of all properties of this {@link Fanta} instance including unrecognized properties.
+     *
+     * @return The map of all properties
+     */
+    @JsonIgnore
+    @Nonnull
+    public Map<String, Object> toMap()
+    {
+        final Map<String, Object> declaredFields = new LinkedHashMap<>(cloudSdkCustomFields);
+        if( sodaType != null )
+            declaredFields.put("sodaType", sodaType);
+        if( color != null )
+            declaredFields.put("color", color);
+        if( flavor != null )
+            declaredFields.put("flavor", flavor);
+        return declaredFields;
     }
 
     /**