+++ datamodel/openapi/openapi-api-sample/src/main/java/com/sap/cloud/sdk/datamodel/openapi/sample/model/Foo.java
@@ -208,6 +208,7 @@ public class Foo implements OneOfWithEnumDiscriminator
     /**
      * Get the value of an unrecognizable property of this {@link Foo} instance.
      *
+     * @deprecated Use {@link #toMap()} instead.
      * @param name
      *            The name of the property
      * @return The value of the property
@@ -215,6 +216,7 @@ public class Foo implements OneOfWithEnumDiscriminator
      *             If no property with the given name could be found.
      */
     @Nullable
+    @Deprecated
     public Object getCustomField( @Nonnull final String name )
         throws NoSuchElementException
     {
@@ -222,6 +224,23 @@ public class Foo implements OneOfWithEnumDiscriminator
             throw new NoSuchElementException("Foo has no field with name '" + name + "'.");
         }
         return cloudSdkCustomFields.get(name);
+    }
+
+    /**
+     * Get the value of all properties of this {@link Foo} instance including unrecognized properties.
+     *
+     * @return The map of all properties
+     */
+    @JsonIgnore
+    @Nonnull
+    public Map<String, Object> toMap()
+    {
+        final Map<String, Object> declaredFields = new LinkedHashMap<>(cloudSdkCustomFields);
+        if( foo != null )
+            declaredFields.put("foo", foo);
+        if( disc != null )
+            declaredFields.put("disc", disc);
+        return declaredFields;
     }
 
     /**