+++ datamodel/openapi/openapi-api-sample/src/main/java/com/sap/cloud/sdk/datamodel/openapi/sample/model/Order.java
@@ -257,6 +257,7 @@ public class Order
     /**
      * Get the value of an unrecognizable property of this {@link Order} instance.
      *
+     * @deprecated Use {@link #toMap()} instead.
      * @param name
      *            The name of the property
      * @return The value of the property
@@ -264,6 +265,7 @@ public class Order
      *             If no property with the given name could be found.
      */
     @Nullable
+    @Deprecated
     public Object getCustomField( @Nonnull final String name )
         throws NoSuchElementException
     {
@@ -271,6 +273,29 @@ public class Order
             throw new NoSuchElementException("Order has no field with name '" + name + "'.");
         }
         return cloudSdkCustomFields.get(name);
+    }
+
+    /**
+     * Get the value of all properties of this {@link Order} instance including unrecognized properties.
+     *
+     * @return The map of all properties
+     */
+    @JsonIgnore
+    @Nonnull
+    public Map<String, Object> toMap()
+    {
+        final Map<String, Object> declaredFields = new LinkedHashMap<>(cloudSdkCustomFields);
+        if( productId != null )
+            declaredFields.put("productId", productId);
+        if( quantity != null )
+            declaredFields.put("quantity", quantity);
+        if( totalPrice != null )
+            declaredFields.put("totalPrice", totalPrice);
+        if( typelessProperty != null )
+            declaredFields.put("typelessProperty", typelessProperty);
+        if( nullableProperty != null )
+            declaredFields.put("nullableProperty", nullableProperty);
+        return declaredFields;
     }
 
     /**