+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/api-class-vendor-extension-json/output/com/sap/cloud/sdk/services/apiclassvendorextension/model/NewSoda.java
@@ -250,16 +250,37 @@ public class NewSoda
 
   /**
    * Get the value of an unrecognizable property of this {@link NewSoda} instance.
+   * @deprecated Use {@link #toMap()} instead.
    * @param name  The name of the property
    * @return The value of the property
    * @throws NoSuchElementException  If no property with the given name could be found.
    */
   @Nullable
+  @Deprecated
   public Object getCustomField( @Nonnull final String name ) throws NoSuchElementException {
     if( !cloudSdkCustomFields.containsKey(name) ) {
         throw new NoSuchElementException("NewSoda has no field with name '" + name + "'.");
     }
     return cloudSdkCustomFields.get(name);
+  }
+
+  /**
+   * Get the value of all properties of this {@link NewSoda} instance including unrecognized properties.
+   *
+   * @return The map of all properties
+   */
+  @JsonIgnore
+  @Nonnull
+  public Map<String, Object> toMap()
+  {
+    final Map<String, Object> declaredFields = new LinkedHashMap<>(cloudSdkCustomFields);
+    if( name != null ) declaredFields.put("name", name);
+    if( brand != null ) declaredFields.put("brand", brand);
+    if( zero != null ) declaredFields.put("zero", zero);
+    if( since != null ) declaredFields.put("since", since);
+    if( flavor != null ) declaredFields.put("flavor", flavor);
+    if( price != null ) declaredFields.put("price", price);
+    return declaredFields;
   }
 
   /**